feat: server factory interfaces extracted
