+++ module/JepRiaShowcase/App/gwt/src/resources/com/technology/jep/jepriashowcase/actuator/version.json
@@ -1,1 +1,1 @@
-{"library":{"Jepria":" 12.1.0"},"compile":{"time_stamp":"2020.12.07 17:31:52","user_name":"SavenkovAV","UUID":"dcc07c4e-70db-448d-8c3e-9bb160957de4","host_name":"MSK-FLEX-D3"},"svn":{"tag_version":"App","repo_name":"jep","module_name":"JepRiaShowcase","revision":""}}
+{"library":{"Jepria":" 12.1.0"},"compile":{"time_stamp":"2020.12.07 19:12:06","user_name":"SavenkovAV","UUID":"fce9d746-e5ed-4887-b519-48a4b043f4f5","host_name":"MSK-FLEX-D3"},"svn":{"tag_version":"App","repo_name":"jep","module_name":"JepRiaShowcase","revision":""}}
