+++ module/JepRiaShowcase/App/service-rest/src/main/java/com/technology/jep/jepriashowcase/featureprocess/FeatureProcessServerFactory.java
@@ -1,28 +1,12 @@
 package com.technology.jep.jepriashowcase.featureprocess;
 
-import com.technology.jep.jepriashowcase.featureprocess.dao.FeatureProcessDao;
-import com.technology.jep.jepriashowcase.featureprocess.dao.FeatureProcessDaoImpl;
-import org.jepria.server.ServerFactory;
 import org.jepria.server.service.rest.EntityService;
-import org.jepria.server.service.rest.EntityServiceImpl;
+import org.jepria.server.service.rest.SearchService;
 
-import javax.inject.Inject;
+import javax.servlet.http.HttpSession;
+import java.util.function.Supplier;
 
-public class FeatureProcessServerFactory extends ServerFactory<FeatureProcessDao> {
-
-  @Inject
-  private FeatureProcessServerFactory(FeatureProcessDao dao) {
-    super(dao, "jdbc/ITMDS");
-  }
-
-  public FeatureProcessService getService() {
-    return new FeatureProcessService(getDao());
-  }
-
-  /**
-   * @return сервис, воплощающий логику CRUD-операций (create, get-by-id, update, delete)
-   */
-  public EntityService getEntityService() {
-    return new EntityServiceImpl(getDao(), new FeatureProcessRecordDefinition());
-  }
+public interface FeatureProcessServerFactory {
+  public FeatureProcessService getService();
+  public EntityService getEntityService();
 }
