+++ module/JepRiaShowcase/App/service-rest/src/main/java/com/technology/jep/jepriashowcase/featureprocess/rest/FeatureProcessJaxrsAdapter.java
@@ -2,6 +2,7 @@ package com.technology.jep.jepriashowcase.featureprocess.rest;
 
 import com.technology.jep.jepriashowcase.featureprocess.FeatureProcessFieldNames;
 import com.technology.jep.jepriashowcase.featureprocess.FeatureProcessServerFactory;
+import com.technology.jep.jepriashowcase.featureprocess.FeatureProcessServerFactoryImpl;
 import com.technology.jep.jepriashowcase.featureprocess.dto.FeatureProcessCreateDto;
 import com.technology.jep.jepriashowcase.featureprocess.dto.FeatureProcessCreateDtoInternal;
 import com.technology.jep.jepriashowcase.featureprocess.dto.FeatureProcessDto;
