+++ module/JepRiaShowcase/App/service-rest/src/main/java/com/technology/jep/jepriashowcase/main/rest/jersey/ApplicationConfig.java
@@ -6,14 +6,17 @@ import com.technology.jep.jepriashowcase.feature.dao.FeatureDao;
 import com.technology.jep.jepriashowcase.feature.dao.FeatureDaoImpl;
 import com.technology.jep.jepriashowcase.feature.rest.FeatureJaxrsAdapter;
 import com.technology.jep.jepriashowcase.featureprocess.FeatureProcessServerFactory;
+import com.technology.jep.jepriashowcase.featureprocess.FeatureProcessServerFactoryImpl;
 import com.technology.jep.jepriashowcase.featureprocess.dao.FeatureProcessDao;
 import com.technology.jep.jepriashowcase.featureprocess.dao.FeatureProcessDaoImpl;
 import com.technology.jep.jepriashowcase.featureprocess.rest.FeatureProcessJaxrsAdapter;
 import com.technology.jep.jepriashowcase.goods.GoodsServerFactory;
+import com.technology.jep.jepriashowcase.goods.GoodsServerFactoryImpl;
 import com.technology.jep.jepriashowcase.goods.dao.GoodsDao;
 import com.technology.jep.jepriashowcase.goods.dao.GoodsDaoImpl;
 import com.technology.jep.jepriashowcase.goods.rest.GoodsJaxrsAdapter;
 import org.glassfish.jersey.internal.inject.AbstractBinder;
+import org.glassfish.jersey.process.internal.RequestScoped;
 import org.jepria.compat.shared.exceptions.ApplicationException;
 import org.jepria.server.service.rest.jersey.ApplicationConfigBase;
 
@@ -46,13 +49,13 @@ public class ApplicationConfig extends ApplicationConfigBase {
     register(new AbstractBinder() {
       @Override
       protected void configure() {
-        bind(FeatureDaoImpl.class).to(FeatureDao.class);
+        bind(FeatureDaoImpl.class).to(FeatureDao.class).in(RequestScoped.class);
       }
     });
     register(new AbstractBinder() {
       @Override
       protected void configure() {
-        bind(FeatureServerFactoryImpl.class).to(FeatureServerFactory.class);
+        bind(FeatureServerFactoryImpl.class).to(FeatureServerFactory.class).in(RequestScoped.class);
       }
     });
     register(FeatureJaxrsAdapter.class);
@@ -62,13 +65,13 @@ public class ApplicationConfig extends ApplicationConfigBase {
     register(new AbstractBinder() {
       @Override
       protected void configure() {
-        bind(FeatureProcessDaoImpl.class).to(FeatureProcessDao.class);
+        bind(FeatureProcessDaoImpl.class).to(FeatureProcessDao.class).in(RequestScoped.class);
       }
     });
     register(new AbstractBinder() {
       @Override
       protected void configure() {
-        bind(FeatureProcessServerFactory.class).to(FeatureProcessServerFactory.class);
+        bind(FeatureProcessServerFactoryImpl.class).to(FeatureProcessServerFactory.class).in(RequestScoped.class);
       }
     });
     register(FeatureProcessJaxrsAdapter.class);
@@ -78,13 +81,13 @@ public class ApplicationConfig extends ApplicationConfigBase {
     register(new AbstractBinder() {
       @Override
       protected void configure() {
-        bind(GoodsDaoImpl.class).to(GoodsDao.class);
+        bind(GoodsDaoImpl.class).to(GoodsDao.class).in(RequestScoped.class);
       }
     });
     register(new AbstractBinder() {
       @Override
       protected void configure() {
-        bind(GoodsServerFactory.class).to(GoodsServerFactory.class);
+        bind(GoodsServerFactoryImpl.class).to(GoodsServerFactory.class).in(RequestScoped.class);
       }
     });
     register(GoodsJaxrsAdapter.class);
